requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-dotenv==1.0.0
tld==0.13.1
transformers==4.35.2
torch==2.1.1
numpy==1.26.2
aiohttp==3.9.1
concurrent-futures==3.0.5
python-whois==0.8.0
pyOpenSSL==23.2.0
cryptography==41.0.4
validators==0.22.0
python-Levenshtein==0.23.0
matplotlib==3.8.2
reportlab==4.0.8
python-magic==0.4.27
hashlib==20081119
python-dateutil==2.8.2